[Fase 2] [REFACTOR] Precisión numérica, DTOs y manejo robusto de fechas - Reemplaza float por Decimal en Reading, TariffCalculator y ReadingService   para evitar errores de redondeo en cálculos monetarios. - Introduce DTO NewReading para mejorar la interfaz del servicio de lecturas. - Mejora el manejo de fechas en SQLiteReadingRepository (uso consistente de datetime). - Actualiza todos los tests para usar Decimal y nuevos DTOs. - Aplica redondeo explícito (ROUND_HALF_UP) en cálculos de tarifa. - Mantiene compatibilidad con float en entrada (conversión segura a Decimal).

+++ app/domain/entities/reading.py
@@ -1,15 +1,15 @@
 from __future__ import annotations
 from dataclasses import dataclass
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
-
 
 @dataclass
 class Reading:
     id: Optional[int]
     user_id: int
-    lectura_actual: float
-    lectura_anterior: float
-    consumo: float
-    costo: float
+    lectura_actual: Decimal
+    lectura_anterior: Decimal
+    consumo: Decimal
+    costo: Decimal
     fecha: datetime | None = None
